moved tools and splitted topAnalysis.h
